Added physics bodies to engine
I added a new PhysicsBody class. It's the same as Character, except it takes three more parameters (gravity, mass, material). It automatically applies gravity. You can apply forces to it via PhysicsBody.applyForces(), and this will move the body. The force decelerates depending on the mass. You can select one of three strings for material: "b" (bouncy), "n" (normal), and "s" (sticky). The bounce function for this class doesn't take parameters (it uses the body's velocity). On material "b", it reverses the velocities. On material "n", it reverses and modifies by the body's mass. With "s", it just sticks. PHYSICSBODY IS STILL IN A VERY EARLY STATE. DO NOT USE IT EXCEPT FOR SIMPLE OBJECTS OR TESTS. CHANGES TO COME.